[cmd/gobox-client] separate event sources into individual files
